Static Typing Everything
Static Typing as much as I possibly can, since this code was written long before I had any idea what static typing is.
This caused the JSON handler to throw some warnings. I am not eager to poke around with the save system so I'm leaving it be for now.